apps/matcher_backend_ai/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
sentence-transformers
pinecone-client[grpc]
pinecone
